--- src/components/Layout/ProjectPageLayout.css
@@ -1,6 +1,0 @@
-.matter-scene {
-  width: 100%;
-  height: 100%;
-  background-color: #fadf0b;
-  overflow: hidden;
-}
