+++ src/components/MatterComponent/MatterComponent.css
@@ -1,0 +1,6 @@
+.matter-scene {
+  width: 100%;
+  height: 100%;
+  background-color: #fadf0b;
+  overflow: hidden;
+}
